requirements-dev.txt
-r requirements.txt
pytest==8.4.1
pytest-django==4.11.1
ruff==0.12.7
